+++ SRP/ControlRoom/Modules/Install/CreateSchema.sql
@@ -336,6 +336,11 @@ INNER JOIN (
 				FROM PatronPoints pp
 				WHERE pp.PID = pt.PID
 				), 0) AS Points,
+		isnull((
+				SELECT isnull(SUM(isnull(NumPoints, 0)), 0)
+				FROM PatronPoints pp
+				WHERE pp.PID = pt.PID AND pp.AwardReasonCd = 0
+				), 0) AS ReadingPoints,
 		pt.TenID
 	FROM Patron pt
 	WHERE pt.PID = @PID
@@ -359,7 +364,7 @@ INNER JOIN (
 	AND (award.NumPoints <= patron.Points)
 	AND (
 		TotalGoal < 1
-		OR award.GoalPercent <= (patron.points * 100) / TotalGoal
+		OR award.GoalPercent <= (patron.ReadingPoints * 100) / TotalGoal
 		)
 	AND (
 		BadgeList = ''
@@ -398,6 +403,11 @@ INNER JOIN (
 				FROM PatronPoints pp
 				WHERE pp.PID = pt.PID
 				), 0) AS Points,
+		isnull((
+				SELECT isnull(SUM(isnull(NumPoints, 0)), 0)
+				FROM PatronPoints pp
+				WHERE pp.PID = pt.PID AND pp.AwardReasonCd = 0
+				), 0) AS ReadingPoints,
 		@TenID AS TenID
 	FROM Patron pt
 	WHERE pt.PID = @PID
@@ -421,7 +431,7 @@ INNER JOIN (
 	AND (award.NumPoints <= patron.Points)
 	AND (
 		TotalGoal < 1
-		OR award.GoalPercent <= (patron.points * 100) / TotalGoal
+		OR award.GoalPercent <= (patron.ReadingPoints * 100) / TotalGoal
 		)
 	AND (
 		BadgeList = ''
@@ -8591,6 +8601,38 @@ BEGIN
 		SELECT 0 AS TotalPoints
 END
 GO
+
+
+
+/****** Object:  StoredProcedure [dbo].[app_PatronPoints_GetPatronReadingPoints]    Script Date: 01/05/2015 14:43:23 ******/
+--Create the Insert Proc
+
+SET ANSI_NULLS ON
+GO
+
+SET QUOTED_IDENTIFIER ON
+GO
+
+
+CREATE PROCEDURE [dbo].[app_PatronPoints_GetPatronReadingPoints] (@PID INT)
+AS
+BEGIN
+	IF (
+			EXISTS (
+				SELECT isnull(SUM(isnull(NumPoints, 0)), 0) AS TotalPoints
+				FROM PatronPoints
+				WHERE PID = @PID AND AwardReasonCd = 0
+				)
+			)
+		SELECT isnull(SUM(isnull(NumPoints, 0)), 0) AS TotalPoints
+		FROM PatronPoints
+		WHERE PID = @PID AND AwardReasonCd = 0
+	ELSE
+		SELECT 0 AS TotalPoints
+END
+
+GO
+
 
 /****** Object:  StoredProcedure [dbo].[app_PatronPoints_Insert]    Script Date: 2/4/2016 13:18:40 ******/
 SET ANSI_NULLS ON