+++ SRP/Controls/MyPointsControl.ascx.designer.cs
@@ -40,6 +40,15 @@ namespace GRA.SRP.Controls {
         protected global::System.Web.UI.HtmlControls.HtmlGenericControl divGoalProgress;
         
         /// <summary>
+        /// lblGoal control.
+        /// </summary>
+        /// <remarks>
+        /// Auto-generated field.
+        /// To modify move field declaration from designer file to code-behind file.
+        /// </remarks>
+        protected global::System.Web.UI.WebControls.Label lblGoal;
+        
+        /// <summary>
         /// divGoalProgressBar control.
         /// </summary>
         /// <remarks>